refactor: add connector firewall source boundary
